move common api stuff to pkg/api
