+++ cmd/agent/common/doget.go
@@ -12,7 +12,7 @@ import (
 	"io/ioutil"
 	"net/http"
 
-	"github.com/DataDog/datadog-agent/cmd/agent/api/common"
+	"github.com/DataDog/datadog-agent/pkg/api/util"
 )
 
 // GetClient is a convenience function returning an http client
@@ -35,7 +35,7 @@ func DoGet(c *http.Client, url string) (body []byte, e error) {
 		return body, e
 	}
 	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+common.GetAuthToken())
+	req.Header.Set("Authorization", "Bearer "+util.GetAuthToken())
 
 	r, e := c.Do(req)
 	if e != nil {
@@ -60,7 +60,7 @@ func DoPost(c *http.Client, url string, contentType string, body io.Reader) (res
 		return resp, e
 	}
 	req.Header.Set("Content-Type", contentType)
-	req.Header.Set("Authorization", "Bearer "+common.GetAuthToken())
+	req.Header.Set("Authorization", "Bearer "+util.GetAuthToken())
 
 	r, e := c.Do(req)
 	if e != nil {
