+++ pkg/api/security/security.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-package common
+package security
 
 import (
 	"crypto/rand"
